refactor: rename `isAssignableTo` to `isSubtypeOf`
This allows us to also add an `isSupertypeOf` method.
